+++ Plugins/AccessibilityPlugin/Native/IQuorumAccessibility/IQuorumAccessibility/ListItemProvider.cpp
@@ -219,7 +219,7 @@ IFACEMETHODIMP ListItemProvider::SetValue(LPCWSTR value)
 IFACEMETHODIMP ListItemProvider::get_Value(BSTR* returnValue)
 {
 	std::wstring text = control->GetText();
-	*returnValue = SysAllocStringLen(text.data(), text.size());
+	*returnValue = SysAllocStringLen(text.data(), static_cast<UINT>(text.size()));
 
 	return S_OK;
 }
